+++ src/backend/config/KsnipConfigOptions.h
@@ -75,6 +75,7 @@ public:
 	static QString uploadScriptPathString();
 	static QString confirmBeforeUploadString();
 	static QString uploadScriptCopyOutputToClipboardString();
+	static QString uploadScriptStopOnStdErr();
 	static QString uploadScriptCopyOutputAfterString();
 	static QString uploadScriptCopyOutputBeforeString();
 	static QString globalHotKeysEnabledString();