+++ src/backend/uploader/script/ScriptUploader.h
@@ -51,6 +51,8 @@ private slots:
 	QString parseOutput(const QString &output) const;
 	void writeToConsole(const QString &output) const;
 	UploadStatus mapErrorTypeToStatus(QProcess::ProcessError errorType) const;
+	void handleSuccess();
+	void handleError(const UploadStatus &status, const QString &stdErrOutput);
 };
 
 #endif //KSNIP_SCRIPTUPLOADER_H