+++ src/gui/settingsDialog/ScriptUploaderSettings.cpp
@@ -23,6 +23,7 @@ ScriptUploaderSettings::ScriptUploaderSettings(KsnipConfig *ksnipConfig) :
 	mConfig(ksnipConfig),
 	mLayout(new QGridLayout(this)),
 	mCopyOutputToClipboardCheckbox(new QCheckBox(this)),
+	mStopOnStdErrCheckbox(new QCheckBox(this)),
 	mCopyOutputAfterLabel(new QLabel(this)),
 	mCopyOutputBeforeLabel(new QLabel(this)),
 	mScriptPathLabel(new QLabel(this)),
@@ -39,6 +40,7 @@ ScriptUploaderSettings::~ScriptUploaderSettings()
 {
 	delete mLayout;
 	delete mCopyOutputToClipboardCheckbox;
+	delete mStopOnStdErrCheckbox;
 	delete mCopyOutputAfterLabel;
 	delete mCopyOutputBeforeLabel;
 	delete mScriptPathLabel;
@@ -50,6 +52,7 @@ ScriptUploaderSettings::~ScriptUploaderSettings()
 
 void ScriptUploaderSettings::saveSettings()
 {
+	mConfig->setUploadScriptStopOnStdErr(mStopOnStdErrCheckbox->isChecked());
 	mConfig->setUploadScriptCopyOutputToClipboard(mCopyOutputToClipboardCheckbox->isChecked());
 	mConfig->setUploadScriptCopyOutputAfter(mCopyOutputAfterLineEdit->text());
 	mConfig->setUploadScriptCopyOutputBefore(mCopyOutputBeforeLineEdit->text());
@@ -58,6 +61,10 @@ void ScriptUploaderSettings::saveSettings()
 
 void ScriptUploaderSettings::initGui()
 {
+	mStopOnStdErrCheckbox->setText(tr("Stop when upload script writes to StdErr"));
+	mStopOnStdErrCheckbox->setToolTip(tr("Marks the upload as failed when script writes to StdErr.\n"
+									        "Without this setting errors in the script will be unnoticed."));
+
 	mCopyOutputToClipboardCheckbox->setText(tr("Copy script output to clipboard"));
 	connect(mCopyOutputToClipboardCheckbox, &QCheckBox::stateChanged, this, &ScriptUploaderSettings::copyToClipboardChanged);
 
@@ -85,14 +92,15 @@ void ScriptUploaderSettings::initGui()
 
 	mLayout->setAlignment(Qt::AlignTop);
 	mLayout->setColumnMinimumWidth(0, 10);
-	mLayout->addWidget(mCopyOutputToClipboardCheckbox, 0, 0, 1, 3);
-	mLayout->addWidget(mCopyOutputAfterLabel, 1, 1, 1, 1);
-	mLayout->addWidget(mCopyOutputAfterLineEdit, 1, 2, 1, 1);
-	mLayout->addWidget(mCopyOutputBeforeLabel, 2, 1, 1, 1);
-	mLayout->addWidget(mCopyOutputBeforeLineEdit, 2, 2, 1, 1);
-	mLayout->addWidget(mScriptPathLabel, 3, 0, 1, 1);
-	mLayout->addWidget(mUploadScriptPathLineEdit, 3, 1, 1, 2);
-	mLayout->addWidget(mBrowseButton, 3, 3, 1, 1);
+	mLayout->addWidget(mStopOnStdErrCheckbox, 0, 0, 1, 3);
+	mLayout->addWidget(mCopyOutputToClipboardCheckbox, 1, 0, 1, 3);
+	mLayout->addWidget(mCopyOutputAfterLabel, 2, 1, 1, 1);
+	mLayout->addWidget(mCopyOutputAfterLineEdit, 2, 2, 1, 1);
+	mLayout->addWidget(mCopyOutputBeforeLabel, 3, 1, 1, 1);
+	mLayout->addWidget(mCopyOutputBeforeLineEdit, 3, 2, 1, 1);
+	mLayout->addWidget(mScriptPathLabel, 4, 0, 1, 1);
+	mLayout->addWidget(mUploadScriptPathLineEdit, 4, 1, 1, 2);
+	mLayout->addWidget(mBrowseButton, 4, 3, 1, 1);
 
 	setTitle(tr("Script Uploader"));
 	setLayout(mLayout);
@@ -100,6 +108,7 @@ void ScriptUploaderSettings::initGui()
 
 void ScriptUploaderSettings::loadConfig()
 {
+	mStopOnStdErrCheckbox->setChecked(mConfig->uploadScriptStopOnStdErr());
 	mCopyOutputToClipboardCheckbox->setChecked(mConfig->uploadScriptCopyOutputToClipboard());
 	mCopyOutputAfterLineEdit->setText(mConfig->uploadScriptCopyOutputAfter());
 	mCopyOutputBeforeLineEdit->setText(mConfig->uploadScriptCopyOutputBefore());